[fuzz] Split health_checker_impl_test for use in fuzzing (#12908)
* Refactored health_checker_impl_test for use in fuzzing
